move stuff out of the test harness
